pyproject.toml
[project]
name = "oracle-ddl-rag-mcp"
version = "0.1.0"
description = "Oracle DDL RAG MCP Server for AI coding assistants"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.2.0",
    "chromadb>=0.5.0",
    "oracledb>=2.0.0",
    "networkx>=3.0",
    "sentence-transformers>=3.0.0",
    "pydantic>=2.0",
    "sqlalchemy>=2.0",
    "pyyaml>=6.0",
    "openai>=1.0.0",
]

[project.scripts]
oracle-ddl-mcp = "oracle_ddl_rag.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/oracle_ddl_rag"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.5.0",
]
